requirements.txt
requests
python-dotenv
statistics
